refactor(agenda): remove deprecated aliases
Closes: #3375
