+++ packages/third-parties/agenda/src/index.ts
@@ -2,9 +2,9 @@
  * @file Automatically generated by @tsed/barrels.
  */
 export * from "./constants/constants.js";
-export * from "./decorators/agenda.js";
 export * from "./decorators/define.js";
 export * from "./decorators/every.js";
+export * from "./decorators/jobController.js";
 export * from "./interfaces/AgendaStore.js";
 export * from "./interfaces/interfaces.js";
 export * from "./services/AgendaService.js";
